# docs/frontend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/files/upload": {
      "post": {
        "operationId": "FileGatewayController_uploadFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "folder": {
                    "type": "string",
                    "example": "uploads"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload file to MinIO (multipart/form-data)",
        "tags": [
          "File"
        ]
      }
    },
    "/files/{key}": {
      "get": {
        "operationId": "FileGatewayController_getFileUrl",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "description": "Object key in MinIO",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expires_in",
            "required": false,
            "in": "query",
            "schema": {
              "example": 3600,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get signed URL for file key",
        "tags": [
          "File"
        ]
      },
      "delete": {
        "operationId": "FileGatewayController_deleteFile",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "description": "Object key in MinIO",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete file by key",
        "tags": [
          "File"
        ]
      }
    },
    "/files/view/{key}": {
      "get": {
        "operationId": "FileGatewayController_viewFile",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "description": "Object key in MinIO",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Public redirect to file URL for image/file viewing",
        "tags": [
          "File"
        ]
      }
    },
    "/files/qr": {
      "post": {
        "operationId": "FileGatewayController_generateQr",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateQrRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate QR and upload to MinIO",
        "tags": [
          "File"
        ]
      }
    },
    "/files/pdf": {
      "post": {
        "operationId": "FileGatewayController_generatePdf",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneratePdfRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate PDF and upload to MinIO",
        "tags": [
          "File"
        ]
      }
    },
...
    "/analytics/dashboard": {
      "get": {
        "operationId": "AnalyticsGatewayController_getDashboard",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Dashboard statistics by requester role",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/analytics/revenue": {
      "get": {
        "operationId": "AnalyticsGatewayController_getRevenue",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "daily",
                "weekly",
                "monthly",
                "yearly"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revenue stats by period",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/analytics/kpi": {
      "get": {
        "operationId": "AnalyticsGatewayController_getKpi",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "KPI stats report",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/analytics/reports/orders": {
      "get": {
        "operationId": "AnalyticsGatewayController_getOrderReport",
        "parameters": [
          {
            "name": "fromDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Order report",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/analytics/reports/finance": {
      "get": {
        "operationId": "AnalyticsGatewayController_getFinanceReport",
        "parameters": [
          {
            "name": "fromDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Finance report",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/analytics/reports/couriers": {
      "get": {
        "operationId": "AnalyticsGatewayController_getCourierReport",
        "parameters": [
          {
            "name": "fromDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Courier report",
        "tags": [
          "Analytics"
        ]
      }
    },
...
  },
  "info": {
    "title": "Elchi API Gateway",
    "description": "API Gateway docs for all microservice routes",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "GenerateQrRequestDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "ORDER-12345"
          },
          "file_name": {
            "type": "string",
            "example": "order-12345-qr.png"
          },
          "prefix": {
            "type": "string",
            "example": "ORD-",
            "enum": [
              "BTB-",
              "BTR-",
              "ORD-"
            ],
            "description": "QR matni oldiga qo‘shiladigan prefix"
          },
          "folder": {
            "type": "string",
            "example": "qr"
          }
        },
        "required": [
          "text"
        ]
      },
      "GeneratePdfRequestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Invoice #12345"
          },
          "content": {
            "type": "string",
            "example": "Customer: Ali Valiyev\nAmount: 120000\nStatus: paid"
          },
          "file_name": {
            "type": "string",
            "example": "invoice-12345.pdf"
          },
          "folder": {
            "type": "string",
            "example": "pdf"
          }
        },
        "required": [
          "content"
        ]
      },
...
    }
  }
}